client/requirements.txt
Flask
grpcio == 1.54.2
grpcio-tools == 1.54.2